upgrade to grails 3.0.x
