+++ .gitignore
@@ -14,3 +14,6 @@ target-eclipse
 *sublime*
 *.zip
 plugin.xml
+.idea
+build
+*.iml
