--- application.properties
@@ -1,4 +1,0 @@
-#Grails Metadata file
-#Sat Nov 09 16:07:44 CET 2013
-app.grails.version=2.1.5
-app.name=slug-generator
