+++ gradle.properties
@@ -1,0 +1,2 @@
+grailsVersion=3.1.8
+gradleWrapperVersion=2.13
