+++ grails-app/controllers/grails/slug/generator/UrlMappings.groovy
@@ -1,0 +1,16 @@
+package grails.slug.generator
+
+class UrlMappings {
+
+    static mappings = {
+        "/$controller/$action?/$id?(.$format)?"{
+            constraints {
+                // apply constraints here
+            }
+        }
+
+        "/"(view:"/index")
+        "500"(view:'/error')
+        "404"(view:'/notFound')
+    }
+}
