--- grails-app/domain/grails/plugins/DummyDomainClass.groovy
@@ -1,9 +1,0 @@
-package grails.plugins
-
-class DummyDomainClass {
-    
-    String name
-
-    static constraints = {
-    }
-}
