+++ src/main/groovy/grails/slug/generator/GrailsSlugGeneratorGrailsPlugin.groovy
@@ -1,0 +1,77 @@
+package grails.slug.generator
+
+import grails.plugins.*
+
+class GrailsSlugGeneratorGrailsPlugin extends Plugin {
+
+    // the version or versions of Grails the plugin is designed for
+    def grailsVersion = "3.0.0 > *"
+    // resources that are excluded from plugin packaging
+    def pluginExcludes = [
+            "grails-app/views/error.gsp",
+            "grails-app/domain/**",
+            "grails-app/controllers/**",
+            "grails-app/taglib/**",
+            "grails-app/i18n/**",
+            "web-app/**"
+    ]
+
+    // TODO Fill in these fields
+    def title = "Grails Slug Generator" // Headline display name of the plugin
+    def author = "Iván López"
+    def authorEmail = "lopez.ivan@gmail.com"
+    def description = '''\
+This plugin generates unique slugs for String properties. Its main use case is to generate unique and "nice" URLs for accessing domain objects.
+
+For example, instead of accessing a user profile by its id (http://www.domain.com/user/25) you can generate an unique URL based on the user's name: http://www.domain.com/user/ivan-lopez (in case the user name is Iván López).
+'''
+    def profiles = ['web']
+
+    // URL to the plugin's documentation
+    def documentation = "https://github.com/lmivan/grails-slug-generator/blob/master/README.md"
+
+    // Extra (optional) plugin metadata
+
+    // License: one of 'APACHE', 'GPL2', 'GPL3'
+    def license = "APACHE"
+
+    // Details of company behind the plugin (if there is one)
+//    def organization = [ name: "My Company", url: "http://www.my-company.com/" ]
+
+    // Any additional developers beyond the author specified above.
+//    def developers = [ [ name: "Joe Bloggs", email: "joe@bloggs.net" ]]
+
+    // Location of the plugin's issue tracker.
+//    def issueManagement = [ system: "JIRA", url: "http://jira.grails.org/browse/GPMYPLUGIN" ]
+
+    // Online location of the plugin's browseable source code.
+//    def scm = [ url: "http://svn.codehaus.org/grails-plugins/" ]
+
+    Closure doWithSpring() { {->
+            // TODO Implement runtime spring config (optional)
+        }
+    }
+
+    void doWithDynamicMethods() {
+        // TODO Implement registering dynamic methods to classes (optional)
+    }
+
+    void doWithApplicationContext() {
+        // TODO Implement post initialization spring config (optional)
+    }
+
+    void onChange(Map<String, Object> event) {
+        // TODO Implement code that is executed when any artefact that this plugin is
+        // watching is modified and reloaded. The event contains: event.source,
+        // event.application, event.manager, event.ctx, and event.plugin.
+    }
+
+    void onConfigChange(Map<String, Object> event) {
+        // TODO Implement code that is executed when the project configuration changes.
+        // The event is the same as for 'onChange'.
+    }
+
+    void onShutdown(Map<String, Object> event) {
+        // TODO Implement code that is executed when the application shuts down (optional)
+    }
+}
